requirements.txt
# چارچوب اصلی
Django==5.0.6
djangorestframework==3.15.0

# تشخیص چهره
dlib==19.24.1
face-recognition==1.3.0
git+https://github.com/ageitgey/face_recognition_models  # اضافه شده
numpy==1.26.3
Pillow==10.2.0

# برای کار با داده‌های base64
python-magic==0.4.27

# برای مدیریت CORS در صورت نیاز
django-cors-headers==4.3.1

# پشتیبانی از دیتابیس PostgreSQL (اختیاری)
psycopg2-binary==2.9.9

# برای محیط تولید (اختیاری)
gunicorn==21.2.0

# محیط توسعه و تست (اختیاری)
django-debug-toolbar==4.3.0
pytest==7.4.4
pytest-django==4.7.0